Support for playing the alarm on the H100 hub
I used a pattern similar to the `set()` API which I felt was appropriate
since you might prefer to use the configured values for some settings.

I've added `stop_alarm` and a method to get a list of available alarm types
(ringtones), `get_supported_alarm_type_list`.

I discovered a `Mute` volume option. This isn't available in the tapo app,
but using it did make the app bell animation run but no d from the hub.

I tried putting through some other random string in the `alarm_volume`
parameter, and it used the configured volume, so apparently `Mute` is a
valid value but I am not sure what use it is.
I've included this in the PR but can remove it if preferred.