feat(tapo): Add discovery support
Introduced support for discovering all Tapo devices on the local network
via the `discover_devices` method in `ApiClient`. This enables
querying devices even when their IP addresses are dynamic or unknown.
However, the process is slower, as it involves scanning the
entire network and awaiting responses from each device to the
discovery request.